Split the lengthy lang to ecos file into a dir
